refactor(data-model): pre-extract ServiceIndexStats to foundation (last observability dep)
Move ServiceIndexStats (66-line struct of primitive i64/f32/f64/bool fields +
serde derives) + the IndexStats type alias from src/core/service_types.rs to
proximadb-data-model (foundation). Re-exported from the source so all callers
are unchanged. Redirected observability's 6 IndexStats imports to the foundation.

This resolves ALL remaining outbound crate:: deps in the extractable observability
code (the storage/ sub-module stays in root — 3 deep service refs). Combined with
the prior pre-extractions (graph-model, tier system, TenantStorageUsage,
GraphOperationsPort, PlanOutput dep elimination), observability's core (minus
storage/) now has ZERO crate::* dependencies except proto = foundation.

Ready for the module extraction (src/observability/* minus storage/ → crate).
